Refactor user data handling; move showUserData and postUserData to component.js and remove profile.js
